Indev (#16)
* path of config file in variable for easier testing
added test cases for main
added test case for organiser.is_not_ready()

* path of config file in variable for easier testing
added test cases for main
added test case for organiser.is_not_ready()

* bump

* bump

* Added toggle cmd to toggle features on/off
Added message if not admin for reset and toggle

* added support for toggable features

* Added json config file support to the toggle command so that the command status is saved when bot is OFF

* Added json support to features handling

* Allowed arguments for toggle command

* Allowed arguments for toggle command

* small fix

* bump

* Removed temporarily unit tests for Gatherbot

* bump

* bump

* bump

* bump

* bump

* bump

* bump

* bump

* bump

* bump

* bump

* bump

* Adds the field to json if not already created

* Adds the field to json if not already created

* readme

* readme

* Added premade feature

* turned off mail notifications

* removed useless import

* updated tests

* bump

* bump

* Internalization for English and French DONE for strings and docstrings. Probably messy, but it works
Added !language command to modify during runtime
Also present in json config file for manual modification
if no locale is present in json config file at startup, default is locale of environment. if not english or french, then it's english.

messages.pot to be used for localization.

* Internalization for English and French DONE for strings and docstrings. Probably messy, but it works
Added !language command to modify during runtime
Also present in json config file for manual modification
if no locale is present in json config file at startup, default is locale of environment. if not english or french, then it's english.

messages.pot to be used for localization.

* Internalization for English and French DONE for strings and docstrings. Probably messy, but it works
Added !language command to modify during runtime
Also present in json config file for manual modification
if no locale is present in json config file at startup, default is locale of environment. if not english or french, then it's english.

messages.pot to be used for localization.