Move events out of class so that they can be tested.
The runtime definition of these was pretty horrible and extremely hard
to test. Discord.py wasn't really designed to play with a class like
this - I wonder if I should try to re-write the whole thing without a
bot class.

I had to skip the discord.py event decorator because the partial/coro
interaction loses the name of the function. Hopefully this won't cause
issues when upgrading discord.py in future.